backend/pyproject.toml
[project]
name = "bot-factory-backend"
version = "0.1.0"
description = "Bot Factory Django Backend"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # Core Django
    "Django>=5.0,<6.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.3.0",
    
    # Admin
    "django-unfold>=0.2.0",
    
    # CORS
    "django-cors-headers>=4.3.0",
    
    # Database
    "psycopg2-binary>=2.9.0",
    
    # Environment variables
    "django-environ>=0.11.0",
    "python-dotenv>=1.0.0",
    
    # Filtering
    "django-filter>=23.3",
    
    # AI Integration
    "google-generativeai>=0.3.0",
    
    # Utilities
    "Pillow>=10.0.0",  # For image handling
    
    # Audio processing
    "SpeechRecognition>=3.10.0",  # For audio transcription
    "pydub>=0.25.1",  # For audio format conversion
    
    # File processing
    "PyPDF2>=3.0.0",  # For PDF text extraction
    "python-docx>=1.1.0",  # For DOCX text extraction
    "pytesseract>=0.3.10",  # For OCR (image text extraction)
    "pdf2image>=1.16.3",  # For PDF to image conversion (OCR)
    
    # Google Speech-to-Text API
    "google-cloud-speech>=2.23.0",  # For professional audio transcription
    
    # Encryption
    "cryptography>=41.0.0",  # For encrypting sensitive data (telegram tokens)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.2",
    "black>=23.7.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "factory-boy>=3.3.0",
]

# Note: For Django projects, we use requirements.txt instead of editable install
# This pyproject.toml is mainly for tool configuration (black, ruff, mypy)

# uv configuration
# Note: We don't use uv's project management for Django
# Instead, we use uv pip install with requirements files

[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

